Added WMS custom params and WMS customizer entry point
Pass all Openlayer WMS params to the custom params per default and added also
wms to the customizer so that the user can change any param if he wants to.

This solve the issue for TIME layer as well as for TRANSPARENT=true layer in
openlayer that needs to pass these parameters as well to print in order to have
them printed the same was as they are rendered in ol.